Edited text_visual example, added escape sequences.
